rename png --> svg (generalize)
